Move Layout types to types.ts file
